requirements.txt
# Core Data Processing
pandas
numpy

# Machine Learning & Embeddings
sentence-transformers
faiss-cpu
torch

# Vector Database & Semantic Search
scikit-learn

# API Framework
flask
flask-cors

# Optional LLM Enhancement
google-generativeai

# Environment Management
python-dotenv

# Performance Monitoring (Optional)
matplotlib
seaborn